A quick fix to move the exportAll API calls under /api/admin.
