+++ doc/sphinx-guides/source/admin/metadataexport.rst
@@ -15,9 +15,9 @@ Batch exports through the API
 
 In addition to the automated exports, a Dataverse admin can start a batch job through the API. The following 2 API calls are provided: 
 
-/api/datasets/exportAll?key=...
+/api/admin/metadata/exportAll?key=...
 
-/api/datasets/reExportAll?key=...
+/api/admin/metadata/reExportAll?key=...
 
 The former will attempt to export all the published, local (non-harvested) datasets that haven't been exported yet. 
 The latter will *force* a re-export of every published, local dataset, regardless of whether it has already been exported or not.
